crypto/blake2b: only fuzz non-generic functions if enabled
